+++ src/utils.fs.ts
@@ -1,14 +1,15 @@
-import * as _ from 'lodash'
-import { existsSync, lstatSync, readdirSync, readFileSync } from 'node:fs'
+import { existsSync, lstatSync, readFileSync } from 'node:fs'
 import { normalize } from 'node:path'
+import { opendir } from 'node:fs/promises'
+import * as _ from 'lodash'
 import * as vscode from 'vscode'
 
 export const toFileUri = (path: string) =>
   vscode.Uri.file(normalize(path))
 
-export const getFileContent = (path: vscode.Uri) => {
+export const getFileContent = (uri: vscode.Uri) => {
   try {
-    return readFileSync(path.fsPath, { encoding: 'utf8' })
+    return readFileSync(uri.fsPath, { encoding: 'utf8' })
   } catch (ex) {
     return null
   }
@@ -20,27 +21,32 @@ export const isExistingDirectory = (path: string | vscode.Uri | null) => {
   return existsSync(path) && lstatSync(path).isDirectory()
 }
 
-export const getFolderContents = (uri: vscode.Uri) => {
-  const dirents = readdirSync(uri.fsPath, { withFileTypes: true })
-  const allPaths = dirents.map((dirent) => dirent.isDirectory()
-    ? getFolderContents(vscode.Uri.joinPath(uri, dirent.name))
-    : [{
-      type: 'file',
-      uri: vscode.Uri.joinPath(uri, dirent.name),
-      content: getFileContent(vscode.Uri.joinPath(uri, dirent.name)),
-    }]
-  )
-
-  // TODO: If they have a .pignore/.pigignore file, should also add the { type: 'dir' } entry
-
-  return allPaths.length
-    ? allPaths.flat(Infinity)
-    : [{ type: 'dir', uri, content: null }]
+export async function* getFolderContents(directoryUri: vscode.Uri) {
+  for await (const dirent of await opendir(directoryUri.fsPath)) {
+    const uri = vscode.Uri.joinPath(directoryUri, dirent.name)
+    // If a directory has yieldable items, don't bother yielding the directory itself.
+    // yieldable items are subdirectories and files, excluding .pig.js, .pignore, and .pigignore files.
+    if (dirent.isDirectory()) {
+      let yielded = false
+      const items = getFolderContents(uri)
+      for await (const content of items) {
+        yield content
+        yielded = true
+      }
+      if (!yielded) yield { uri, dirent, content: null }
+    } else if (dirent.isFile() && !['.pig.js', '.pignore', '.pigignore'].includes(dirent.name.toLowerCase())) {
+      yield { 
+        uri, 
+        dirent,
+        get content() { return getFileContent(uri) }
+      }
+    }
+  }
 }
 
-export const getRelativePath = (ancestor: vscode.Uri, descendant: vscode.Uri) => {
-  const afsPath = ancestor.fsPath
-  const dfsPath = descendant.fsPath
+export const getRelativePath = (ancestorUri: vscode.Uri, descendantUri: vscode.Uri) => {
+  const afsPath = ancestorUri.fsPath
+  const dfsPath = descendantUri.fsPath
   const regex = new RegExp(`^${_.escapeRegExp(afsPath)}[/\\\\]`)
   return dfsPath.replace(regex, '')
 }